Fixes an additional bug I found on CONTRIB-4597
https://tracker.moodle.org/browse/CONTRIB-4597

If the fixedstudent flag is set, and the edit mode is turned on, the
rows are off by 2-3 every 50 rows in the grader report. Off by a full row
after the first ~11 rows.

As always, this comes with a disclaimer that I'm not a theme designer. The below fix seems
to work for me, but please use at your discretion.

*Additional commit needed to fix missing semicolon.