#634: Improve icon replacement performance
Load and store the `$icons` array only once and make it accessible for future calls to `replace_moodle_icon`. The array is moved into a static field that is initialised once.

Furthermore, call `self::replace_moodle_icon` once instead of twice. This results in performing (about) half the lookups for replacement icons.