Accessibility: politely Inform the the screen reader on slide change
Hi Gareth,
I am getting an accessibility audit to our system and to the Essential theme.
So here is one tiny and important fix.
Which instructs the screen reader to wait until it finishes reading the current slide's content before starting to read the new slide that already switched. which, btw, I am told... is confusing to the listener and so we made the interval bigger and tried to shorten the text we put inside the sliders.

Also, as an alternative, our accessibility advisor asked us to disable the auto switching of slides. 

Maybe we should put some good advise in the theme settings help? as I have no idea how system administrator with no experience using a screen reader can know about these anomalies.