Accessibility: missing ALT attribute in course summary images
I added the filename as ALT, which seems wrong, but as far as I researched the Moodle code... using Atto for adding an Image, the description is immediately saved as an ALT attribute in the content, but not in the mdl_files DB table. and so, unavailable if we use that image file elsewhere like at the above course summary attached files (and not as links but as images) looks like we need a new accessibility MDL on the tracker. what do you think?